+++ src/main/kotlin/community/whatever/onembackendkotlin/application/UrlShortenService.kt
@@ -1,10 +1,10 @@
 package community.whatever.onembackendkotlin.application
 
-import community.whatever.onembackendkotlin.common.dto.OriginUrlResponse
-import community.whatever.onembackendkotlin.common.dto.ShortenUrlCreateRequest
-import community.whatever.onembackendkotlin.common.dto.ShortenUrlSearchRequest
-import community.whatever.onembackendkotlin.common.dto.ShortenedUrlResponse
-import community.whatever.onembackendkotlin.common.exception.UrlNotFoundException
+import community.whatever.onembackendkotlin.application.dto.OriginUrlResponse
+import community.whatever.onembackendkotlin.application.dto.ShortenUrlCreateRequest
+import community.whatever.onembackendkotlin.application.dto.ShortenUrlSearchRequest
+import community.whatever.onembackendkotlin.application.dto.ShortenedUrlResponse
+import community.whatever.onembackendkotlin.application.exception.UrlNotFoundException
 import community.whatever.onembackendkotlin.domain.ShortenedUrl
 import community.whatever.onembackendkotlin.domain.ShortenedUrlRepository
 import org.springframework.stereotype.Service
