+++ src/main/kotlin/community/whatever/onembackendkotlin/application/dto/UrlShortenRequest.kt
@@ -1,4 +1,4 @@
-package community.whatever.onembackendkotlin.common.dto
+package community.whatever.onembackendkotlin.application.dto
 
 data class ShortenUrlCreateRequest(val originUrl: String)
