+++ src/main/kotlin/community/whatever/onembackendkotlin/application/dto/UrlShortenResponse.kt
@@ -1,4 +1,4 @@
-package community.whatever.onembackendkotlin.common.dto
+package community.whatever.onembackendkotlin.application.dto
 
 data class ShortenedUrlResponse(val shortenedUrl: String)
