+++ src/test/kotlin/community/whatever/onembackendkotlin/presentation/UrlShortenControllerTest.kt
@@ -2,11 +2,11 @@ package community.whatever.onembackendkotlin.presentation
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import community.whatever.onembackendkotlin.application.UrlShortenService
-import community.whatever.onembackendkotlin.common.dto.OriginUrlResponse
-import community.whatever.onembackendkotlin.common.dto.ShortenUrlCreateRequest
-import community.whatever.onembackendkotlin.common.dto.ShortenUrlSearchRequest
-import community.whatever.onembackendkotlin.common.dto.ShortenedUrlResponse
-import community.whatever.onembackendkotlin.common.exception.UrlNotFoundException
+import community.whatever.onembackendkotlin.application.dto.OriginUrlResponse
+import community.whatever.onembackendkotlin.application.dto.ShortenUrlCreateRequest
+import community.whatever.onembackendkotlin.application.dto.ShortenUrlSearchRequest
+import community.whatever.onembackendkotlin.application.dto.ShortenedUrlResponse
+import community.whatever.onembackendkotlin.application.exception.UrlNotFoundException
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.mockito.BDDMockito.given
